refactor: FilterHeader의 네이밍을 SortHeader로 수정했습니다.
